feat: add console app, update doubles to decimals

+++ src/Starbuzz.Domain/Abstract/Beverage.cs
@@ -6,6 +6,6 @@
 
         public virtual string GetDescription() => this.Description;
 
-        public abstract double Cost();
+        public abstract decimal Cost();
     }
 }
